Create and reuse single formatter for all cell views and group header views
